requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
openai>=1.30.0
markdown==3.5.2
httpx>=0.27.0
python-multipart==0.0.6
pymysql==1.1.0
cryptography==42.0.0
